爆発範囲実装

+++ Assets/Saijou/Scripts/PlayerData/PlayerStats.cs
@@ -18,7 +18,7 @@ public class PlayerStats : ScriptableObject
     public int lightningBulletUP = 0; // 雷の弾で感電する敵＋1
     public int chainBulletUP = 0;     // 鎖の弾で拘束する敵＋1
     public int poisonRangeUP = 0;     // 毒の範囲UP
-    public int explosionRangeUP = 0;  // 爆発の範囲UP
+    public float explosionRangeUP = 0;  // 爆発の範囲UP
 
     // 解放した属性弾リスト
     public GameObject[] unlockedElementalBullets;
